Update egui to 0.15
- Replaces `egui_winit_platform` with `egui-winit`
- Renames the example package to avoid crate name collision
