--- examples/egui-winit/src/gui.rs
@@ -1,141 +1,0 @@
-use egui::{ClippedMesh, FontDefinitions};
-use egui_wgpu_backend::{BackendError, RenderPass, ScreenDescriptor};
-use egui_winit_platform::{Platform, PlatformDescriptor};
-use pixels::{wgpu, PixelsContext};
-use std::time::Instant;
-use winit::window::Window;
-
-/// Manages all state required for rendering egui over `Pixels`.
-pub(crate) struct Gui {
-    // State for egui.
-    start_time: Instant,
-    platform: Platform,
-    screen_descriptor: ScreenDescriptor,
-    rpass: RenderPass,
-    paint_jobs: Vec<ClippedMesh>,
-
-    // State for the demo app.
-    window_open: bool,
-}
-
-impl Gui {
-    /// Create egui.
-    pub(crate) fn new(width: u32, height: u32, scale_factor: f64, pixels: &pixels::Pixels) -> Self {
-        let platform = Platform::new(PlatformDescriptor {
-            physical_width: width,
-            physical_height: height,
-            scale_factor,
-            font_definitions: FontDefinitions::default(),
-            style: Default::default(),
-        });
-        let screen_descriptor = ScreenDescriptor {
-            physical_width: width,
-            physical_height: height,
-            scale_factor: scale_factor as f32,
-        };
-        let rpass = RenderPass::new(pixels.device(), pixels.render_texture_format(), 1);
-
-        Self {
-            start_time: Instant::now(),
-            platform,
-            screen_descriptor,
-            rpass,
-            paint_jobs: Vec::new(),
-            window_open: true,
-        }
-    }
-
-    /// Handle input events from the window manager.
-    pub(crate) fn handle_event(&mut self, event: &winit::event::Event<'_, ()>) {
-        self.platform.handle_event(event);
-    }
-
-    /// Resize egui.
-    pub(crate) fn resize(&mut self, width: u32, height: u32) {
-        if width > 0 && height > 0 {
-            self.screen_descriptor.physical_width = width;
-            self.screen_descriptor.physical_height = height;
-        }
-    }
-
-    /// Update scaling factor.
-    pub(crate) fn scale_factor(&mut self, scale_factor: f64) {
-        self.screen_descriptor.scale_factor = scale_factor as f32;
-    }
-
-    /// Prepare egui.
-    pub(crate) fn prepare(&mut self, window: &Window) {
-        self.platform
-            .update_time(self.start_time.elapsed().as_secs_f64());
-
-        // Begin the egui frame.
-        self.platform.begin_frame();
-
-        // Draw the demo application.
-        self.ui(&self.platform.context());
-
-        // End the egui frame and create all paint jobs to prepare for rendering.
-        let (_output, paint_commands) = self.platform.end_frame(Some(window));
-        self.paint_jobs = self.platform.context().tessellate(paint_commands);
-    }
-
-    /// Create the UI using egui.
-    fn ui(&mut self, ctx: &egui::CtxRef) {
-        egui::TopBottomPanel::top("menubar_container").show(ctx, |ui| {
-            egui::menu::bar(ui, |ui| {
-                egui::menu::menu(ui, "File", |ui| {
-                    if ui.button("About...").clicked() {
-                        self.window_open = true;
-                    }
-                })
-            });
-        });
-
-        egui::Window::new("Hello, egui!")
-            .open(&mut self.window_open)
-            .show(ctx, |ui| {
-                ui.label("This example demonstrates using egui with pixels.");
-                ui.label("Made with 💖 in San Francisco!");
-
-                ui.separator();
-
-                ui.horizontal(|ui| {
-                    ui.spacing_mut().item_spacing.x /= 2.0;
-                    ui.label("Learn more about egui at");
-                    ui.hyperlink("https://docs.rs/egui");
-                });
-            });
-    }
-
-    /// Render egui.
-    pub(crate) fn render(
-        &mut self,
-        encoder: &mut wgpu::CommandEncoder,
-        render_target: &wgpu::TextureView,
-        context: &PixelsContext,
-    ) -> Result<(), BackendError> {
-        // Upload all resources to the GPU.
-        self.rpass.update_texture(
-            &context.device,
-            &context.queue,
-            &self.platform.context().texture(),
-        );
-        self.rpass
-            .update_user_textures(&context.device, &context.queue);
-        self.rpass.update_buffers(
-            &context.device,
-            &context.queue,
-            &self.paint_jobs,
-            &self.screen_descriptor,
-        );
-
-        // Record all render passes.
-        self.rpass.execute(
-            encoder,
-            render_target,
-            &self.paint_jobs,
-            &self.screen_descriptor,
-            None,
-        )
-    }
-}
