+++ examples/minimal-egui/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "egui-winit"
+name = "minimal-egui"
 version = "0.1.0"
 authors = ["Jay Oster <jay@kodewerx.org>"]
 edition = "2018"
@@ -11,9 +11,9 @@ optimize = ["log/release_max_level_warn"]
 default = ["optimize"]
 
 [dependencies]
-egui = "0.14"
-egui_wgpu_backend = "0.13"
-egui_winit_platform = { version = "0.10", features = ["webbrowser"] }
+egui = "0.15"
+egui_wgpu_backend = { git = "https://github.com/hasenbanck/egui_wgpu_backend.git", rev = "961125e7bd2c71c5ead1d61a7ca7ffa8c0d17f48" }
+egui-winit = { version = "0.15", default-features = false, features = ["links"] }
 env_logger = "0.9"
 log = "0.4"
 pixels = { path = "../.." }
