+++ examples/minimal-egui/src/gui.rs
@@ -1,0 +1,145 @@
+use egui::{ClippedMesh, CtxRef};
+use egui_wgpu_backend::{BackendError, RenderPass, ScreenDescriptor};
+use pixels::{wgpu, PixelsContext};
+use winit::window::Window;
+
+/// Manages all state required for rendering egui over `Pixels`.
+pub(crate) struct Framework {
+    // State for egui.
+    egui_ctx: CtxRef,
+    egui_state: egui_winit::State,
+    screen_descriptor: ScreenDescriptor,
+    rpass: RenderPass,
+    paint_jobs: Vec<ClippedMesh>,
+
+    // State for the GUI
+    gui: Gui,
+}
+
+/// Example application state. A real application will need a lot more state than this.
+struct Gui {
+    /// Only show the egui window when true.
+    window_open: bool,
+}
+
+impl Framework {
+    /// Create egui.
+    pub(crate) fn new(width: u32, height: u32, scale_factor: f32, pixels: &pixels::Pixels) -> Self {
+        let egui_ctx = CtxRef::default();
+        let egui_state = egui_winit::State::from_pixels_per_point(scale_factor);
+        let screen_descriptor = ScreenDescriptor {
+            physical_width: width,
+            physical_height: height,
+            scale_factor,
+        };
+        let rpass = RenderPass::new(pixels.device(), pixels.render_texture_format(), 1);
+        let gui = Gui::new();
+
+        Self {
+            egui_ctx,
+            egui_state,
+            screen_descriptor,
+            rpass,
+            paint_jobs: Vec::new(),
+            gui,
+        }
+    }
+
+    /// Handle input events from the window manager.
+    pub(crate) fn handle_event(&mut self, event: &winit::event::WindowEvent) {
+        self.egui_state.on_event(&self.egui_ctx, event);
+    }
+
+    /// Resize egui.
+    pub(crate) fn resize(&mut self, width: u32, height: u32) {
+        if width > 0 && height > 0 {
+            self.screen_descriptor.physical_width = width;
+            self.screen_descriptor.physical_height = height;
+        }
+    }
+
+    /// Update scaling factor.
+    pub(crate) fn scale_factor(&mut self, scale_factor: f64) {
+        self.screen_descriptor.scale_factor = scale_factor as f32;
+    }
+
+    /// Prepare egui.
+    pub(crate) fn prepare(&mut self, window: &Window) {
+        // Begin the egui frame.
+        let raw_input = self.egui_state.take_egui_input(window);
+        self.egui_ctx.begin_frame(raw_input);
+
+        // Draw the demo application.
+        self.gui.ui(&self.egui_ctx);
+
+        // End the egui frame and create all paint jobs to prepare for rendering.
+        let (output, paint_commands) = self.egui_ctx.end_frame();
+        self.egui_state
+            .handle_output(window, &self.egui_ctx, output);
+        self.paint_jobs = self.egui_ctx.tessellate(paint_commands);
+    }
+
+    /// Render egui.
+    pub(crate) fn render(
+        &mut self,
+        encoder: &mut wgpu::CommandEncoder,
+        render_target: &wgpu::TextureView,
+        context: &PixelsContext,
+    ) -> Result<(), BackendError> {
+        // Upload all resources to the GPU.
+        self.rpass
+            .update_texture(&context.device, &context.queue, &self.egui_ctx.texture());
+        self.rpass
+            .update_user_textures(&context.device, &context.queue);
+        self.rpass.update_buffers(
+            &context.device,
+            &context.queue,
+            &self.paint_jobs,
+            &self.screen_descriptor,
+        );
+
+        // Record all render passes.
+        self.rpass.execute(
+            encoder,
+            render_target,
+            &self.paint_jobs,
+            &self.screen_descriptor,
+            None,
+        )
+    }
+}
+
+impl Gui {
+    /// Create a `Gui`.
+    fn new() -> Self {
+        Self { window_open: true }
+    }
+
+    /// Create the UI using egui.
+    fn ui(&mut self, ctx: &CtxRef) {
+        egui::TopBottomPanel::top("menubar_container").show(ctx, |ui| {
+            egui::menu::bar(ui, |ui| {
+                egui::menu::menu(ui, "File", |ui| {
+                    if ui.button("About...").clicked() {
+                        self.window_open = true;
+                    }
+                })
+            });
+        });
+
+        egui::Window::new("Hello, egui!")
+            .open(&mut self.window_open)
+            .show(ctx, |ui| {
+                ui.label("This example demonstrates using egui with pixels.");
+                ui.label("Made with 💖 in San Francisco!");
+
+                ui.separator();
+
+                ui.horizontal(|ui| {
+                    ui.spacing_mut().item_spacing.x /= 2.0;
+                    ui.label("Learn more about egui at");
+                    ui.hyperlink("https://docs.rs/egui");
+                });
+            });
+    }
+}
